fixup: move cases which verify create/drop/if not exists logic to TableEnvironmentTest
